API/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
Flask-RESTful==0.3.8
Flask-Cors==3.0.8
mysql-connector-python==8.0.21
requests==2.20.1
urllib3==1.24.2
DateTime==4.3
elasticsearch==7.13.1
python-dotenv==0.18.0
forex-python==1.6